refactor: delete site_builder_anonymizer
