+++ great_expectations/core/usage_statistics/anonymizers/data_docs_site_anonymizer.py
@@ -1,7 +1,4 @@
 from great_expectations.core.usage_statistics.anonymizers.anonymizer import Anonymizer
-from great_expectations.core.usage_statistics.anonymizers.site_builder_anonymizer import (
-    SiteBuilderAnonymizer,
-)
 from great_expectations.core.usage_statistics.anonymizers.store_backend_anonymizer import (
     StoreBackendAnonymizer,
 )
@@ -10,7 +7,6 @@ from great_expectations.core.usage_statistics.anonymizers.store_backend_anonymiz
 class DataDocsSiteAnonymizer(Anonymizer):
     def __init__(self, salt=None):
         super().__init__(salt=salt)
-        self._site_builder_anonymizer = SiteBuilderAnonymizer(salt=salt)
         self._store_backend_anonymizer = StoreBackendAnonymizer(salt=salt)
 
     def anonymize_data_docs_site_info(self, site_name, site_config):
@@ -20,10 +16,8 @@ class DataDocsSiteAnonymizer(Anonymizer):
                 "module_name"
             ] = "great_expectations.render.renderer.site_builder"
 
-        anonymized_info_dict = (
-            self._site_builder_anonymizer.anonymize_site_builder_info(
-                site_builder_config=site_config,
-            )
+        anonymized_info_dict = self.anonymize_site_builder_info(
+            site_builder_config=site_config,
         )
         anonymized_info_dict["anonymized_name"] = self.anonymize(site_name)
 
@@ -34,10 +28,8 @@ class DataDocsSiteAnonymizer(Anonymizer):
             store_backend_object_config=store_backend_config
         )
         site_index_builder_config = site_config.get("site_index_builder")
-        anonymized_site_index_builder = (
-            self._site_builder_anonymizer.anonymize_site_builder_info(
-                site_builder_config=site_index_builder_config
-            )
+        anonymized_site_index_builder = self.anonymize_site_builder_info(
+            site_builder_config=site_index_builder_config
         )
         # Note AJB-20201218 show_cta_footer was removed in v 0.9.9 via PR #1249
         if "show_cta_footer" in site_index_builder_config:
